204 for all error conditions on userimage
If an image is requested that doesn't exist, userimage will now return
HTTP 204 No Content, the same as if you didn't have permission to view
it. A crude sort of "please don't hack us" check for the QUERY_STRING
that does a greedy strip of /* and ;* should stop any obvious attacks
through this command, although I'm still a little uneasy putting a bash
script on the internet like that.